Add crop-to-fit scaling feature.
Input file is cropped and printed. We don't do scaling in crop-to-fit.
In pdftopdf filter, a check is created so that fillprint and cropfit are
applied only once. Fixes #65. For crop-to-fit, orientation should be ignored.
TODO: Ignore orientation in imageto... filters.

"-o fill" option is now equivalent to "print-scaling=fill".